FILE: moves ApiService into internal folder
